move free allocation tests to flaky scenarios
